Fold datetimeutil into util
There isn't any reason to have datetimeutil separate. This folds it in.
