--- antenna/datetimeutil.py
@@ -1,33 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-# This is based on socorrolib's socorrolib/lib/datetimeutil.py module.
-
-import datetime
-
-import isodate
-
-UTC = isodate.UTC
-
-
-def utc_now():
-    """Return a timezone aware datetime instance in UTC timezone
-
-    This funciton is mainly for convenience. Compare:
-
-        >>> from antenna.datetimeutil import utc_now
-        >>> utc_now()
-        datetime.datetime(2012, 1, 5, 16, 42, 13, 639834,
-          tzinfo=<isodate.tzinfo.Utc object at 0x101475210>)
-
-    Versus:
-
-        >>> import datetime
-        >>> from antenna.datetimeutil import UTC
-        >>> datetime.datetime.now(UTC)
-        datetime.datetime(2012, 1, 5, 16, 42, 13, 639834,
-          tzinfo=<isodate.tzinfo.Utc object at 0x101475210>)
-
-    """
-    return datetime.datetime.now(UTC)
